requirements.txt
matplotlib~=3.8.4
ml_collections==0.1.1
numpy~=1.26.3
opencv_contrib_python==4.7.0.72
pandas~=2.2.2
PyYAML~=6.0.1
scikit_image==0.19.3
scikit_learn==1.0.2
scipy~=1.13.0
torch_pruning==1.2.0
torchvision~=0.18.0+cu118
torchviz==0.0.2
tqdm~=4.66.4
scikit-learn~=1.4.2
opencv-python~=4.10.0.84
scikit-image~=0.23.2
ml_collections
pandas
scikit-learn
torch~=2.3.0+cu118
albumentations~=1.4.11
pillow~=10.2.0
pyautogui~=0.9.54